+++ tests/test_scf.py
@@ -56,8 +56,8 @@ def test_sic():
 def test_symmetric():
     """Test the symmetry option for H2 dissociation."""
     atoms = Atoms('H2', ((0, 0, 0), (0, 0, 6)), Nspin=2, ecut=1)
-    e_symm = SCF(atoms, symmetric=True).run()
-    e_unsymm = SCF(atoms, symmetric=False).run()
+    e_symm = SCF(atoms, guess='pseudo', symmetric=True).run()
+    e_unsymm = SCF(atoms, guess='pseudo', symmetric=False).run()
     assert e_symm > e_unsymm