+++ CHANGELOG.md
@@ -22,6 +22,7 @@ dev
    - Reduce McCabe code complexity
    - Switch linter from flake8 to Ruff
    - Comply with different linting rules, e.g., use triple-quotes in docstrings
+   - More tests and more coverage
 - Miscellaneous
    - Performance fix by using precomputed values correctly
    - Improve GGA performance